+++ vector3D.h
@@ -5,22 +5,22 @@
 
 struct vector3D
 {
-    double x,y,z;
+    long double x,y,z;
 }; 
 
 void get_vector(struct vector3D *v)
 {
     printf("Imput the x component of the vector\n");
-    scanf("%lf",&v->x);
+    scanf("%Lf",&v->x);
     printf("Imput the y component of the vector\n");
-    scanf("%lf",&v->y);
+    scanf("%Lf",&v->y);
     printf("Imput the z component of the vector\n");
-    scanf("%lf",&v->z);
+    scanf("%Lf",&v->z);
 }
 
 void print_vector(struct vector3D v)
 {
-    printf("(%lf,\t%lf,\t%lf)\n",v.x,v.y,v.z);
+    printf("(%Lf,\t%Lf,\t%Lf)\n",v.x,v.y,v.z);
 }
 
 /*const char *to_string(struct vector3D v)
@@ -38,17 +38,17 @@ struct vector3D resta(struct vector3D v,struct vector3D w)
     return suma(v,(struct vector3D){-w.x,-w.y,-w.z});
 }
 
-double module(struct vector3D v)
+long double module(struct vector3D v)
 {
     return sqrt(v.x*v.x+v.y*v.y+v.z*v.z);
 }
 
-double distancia(struct vector3D v,struct vector3D w)
+long double distancia(struct vector3D v,struct vector3D w)
 {
     return module(resta(v,w));
 }
 
-double scalar_product(struct vector3D v,struct vector3D w)
+long double scalar_product(struct vector3D v,struct vector3D w)
 {
     return v.x*w.x+v.y*w.y+v.z*w.z;
 }
